trade_api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.2
duckduckgo-search==4.1.1
slowapi==0.1.9
pydantic-settings==2.1.0
